012 - till max char
Added to only limit the count to the maximum number of char,
i.e. any char lets say there are 1,2,3,4 cahr words.
if there is no 3 char words but the max char word is 4,
then 3 will be shown as 0/nothing, and if 4 is max char word in the input/file,
so after that if all are 0, means no words with 5,6,7..chars,
then last number will only be till 4 char words.